refactor(test): extract preview test helpers
